int for color step real

+++ Runtime/Shaders/UniversalToonInput.hlsl
@@ -34,7 +34,7 @@ fixed _Set_SystemShadowsToBase;
 float _Tweak_SystemShadowsLevel;
 float _BaseColor_Step;
 float _BaseShade_Feather;
-float _NormalMap_Object_Space_Step;
+int _NormalMap_Object_Space_Step;
 
 
 float4 _Set_1st_ShadePosition_ST;
